<fix>: <알림 토스트 디버깅>
