+++ src/components/common/Toast.tsx
@@ -1,62 +1,23 @@
 import { useEffect, useState } from "react";
 import { useNotificationStore } from "../../store/notification-store";
 import { useNavigate } from "react-router";
-import { NEW_NOTIFICATION } from "@/constants/session-storage";
-import { NotificationType } from "@/types/notification-type";
 
 const Toast = ({
-  newIcon,
   setNewIcon,
+  setNewToast,
+  newMessage,
 }: {
-  newIcon: boolean;
   setNewIcon: React.Dispatch<React.SetStateAction<boolean>>;
+  setNewToast: React.Dispatch<React.SetStateAction<boolean>>;
+  newMessage: string;
 }) => {
   const navigate = useNavigate();
-  const [newToast, setNewToast] = useState(false);
-  const [newMessage, setNewMessage] = useState("");
+  const [toast, setToast] = useState(false);
   const { newNotification } = useNotificationStore();
 
   useEffect(() => {
-    if (newIcon) {
-      const newNotificationJSON = sessionStorage.getItem(NEW_NOTIFICATION);
-      if (newNotificationJSON) {
-        const newNotification = JSON.parse(newNotificationJSON) as NotificationType;
-        setNewMessage(newNotification.content);
-        setNewToast(true);
-
-        const timeoutId = setTimeout(() => {
-          setNewToast(false);
-        }, 5000);
-
-        return () => clearTimeout(timeoutId);
-      }
-    }
-  }, [newIcon]);
-
-  // useEffect(() => {
-  //   if (newNotification) {
-  //     // 중복 렌더링을 피하기 위해 한 번 거름
-  //     const isDuplicate = notificationList.some((noti) => noti.notificationId === newNotification.notificationId);
-  //     if (!isDuplicate) {
-  //       setNotificationList([newNotification, ...notificationList]);
-  //     }
-  //     (async () => {
-  //       await makeToast();
-  //       if (newNotification) {
-  //         setNewMessage(() => newNotification.content);
-  //       }
-  //       setTimeout(() => {
-  //         // 5초 후에 초기화
-  //         setNewToast(() => false);
-  //         setNewNotification(null);
-  //       }, 5000);
-  //     })();
-  //   }
-  // }, [newNotification, notificationList, setNewNotification, setNotificationList]);
-
-  // const makeToast = async () => {
-  //   setNewToast(() => true);
-  // };
+    setToast(() => true);
+  }, [setNewToast]);
 
   const handleClick = () => {
     if (newNotification) {
@@ -70,7 +31,7 @@ const Toast = ({
     <>
       <div
         onClick={() => handleClick()}
-        className={`w-52 h-14 bg-Gray-1 text-Blue-2 text-xs rounded-[10px] p-2 flex justify-center items-center absolute z-50 left-[5%] shadow-card transition-all ${newToast ? "top-20 opacity-100" : "top-14 opacity-0 pointer-events-none"}`}
+        className={`w-52 h-14 bg-Gray-1 text-Blue-2 text-xs rounded-[10px] p-2 flex justify-center items-center absolute z-50 left-[5%] shadow-card transition-all ${toast ? "top-20 opacity-100" : "top-14 opacity-0 pointer-events-none"}`}
       >
         <svg
           xmlns="http://www.w3.org/2000/svg"
